requirements.txt
aiohttp==3.6.3
discord.py==1.5.1
fuzzywuzzy==0.18.0
numpy==1.19.4
opencv-python==4.4.0.46
python-dotenv==0.15.0
validators==0.18.1
python-Levenshtein==0.12.0
